Cleared a TODO.
